+++ ESFramework/UIKit/UIView+ESAdditions.h
@@ -56,12 +56,12 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  * Searches in superviews recursively to find the view which its class is the given viewClass.
  */
-- (nullable UIView *)findSuperviewOf:(Class)viewClass;
+- (nullable UIView *)findSuperviewOfClass:(Class)viewClass;
 
 /**
  * Searches in subviews recursively to find the view which its class is the given viewClass.
  */
-- (nullable UIView *)findSubviewOf:(Class)viewClass;
+- (nullable UIView *)findSubviewOfClass:(Class)viewClass;
 
 /**
  * Returns the UIViewController instance which manages this view.